+++ src/media/tools/movies/namelist.py
@@ -7,73 +7,140 @@ Module to report on all named within a movie
 
 import os
 import argparse
-from media.tools.common import load_media_dev, compile_movies
+from media.tools.common import (
+        load_media_dev, compile_movies, random_sample_list
+        )
 
 
-class NameList():
-    '''Build a structure of names and job roles'''
-    def __init__(self):
-        self.names = {}
+class NameJobTitleMap():
+    '''Crewname object that lists every job role in every title.
+    '''
+    def __init__(self, in_name, in_job, in_title):
+        self.name = in_name
+        self.jobs = {}
+        self.jobs[in_job] = [in_title]
 
-    def add(self, name, job_role, title):
-        '''Add a name, job role, and title to the list'''
-        if name in self.names:
-            if job_role in self.names[name]:
-                self.names[name][job_role].append(title)
-            else:
-                self.names[name][job_role] = [title]
+    def add_job_title(self, in_job, in_title):
+        '''
+        Add a job role and movie title to the name.
+        '''
+        if in_job in self.jobs:
+            self.jobs[in_job].append(in_title)
         else:
-            self.names[name] = {}
-            self.names[name][job_role] = [title]
+            self.jobs[in_job] = [in_title]
 
-    def output(self):
-        '''Return the names, job roles, and titles'''
-        print(f"{'Family Name':20s} {'Given Name':15s} " +
-              f"{'Job Role':20s} {'Title'}")
-        print(f"{'=' * 20} {'=' * 15} {'=' * 20} {'=' * 25}")
-        for nam in sorted(self.names.keys()):
-            for j_role in sorted(self.names[nam].keys()):
-                t_count = 1
-                for title in sorted(self.names[nam][j_role]):
-                    if t_count == 1:
-                        print(f"{nam.family:20s} {nam.given:15s} " +
-                              f"{j_role:20s} {title!s}")
+    @classmethod
+    def header(cls):
+        '''
+        Output a simple header.
+        '''
+        out = f"{'Family Name':20s} {'Given Name':15s} " + \
+              f"{'Job':20s} {'Title':25s}\n" + \
+              f"{'=' * 20} {'=' * 15} {'=' * 20} {'=' * 25}"
+        return out
+
+    def __str__(self):
+        out = ''
+        name_l = 1
+        for job_i in sorted(self.jobs):
+            job_l = 1
+            for title_i in sorted(self.jobs[job_i]):
+                if name_l == 1:
+                    out = f"{self.name.family:20s} {self.name.given:15s} " + \
+                          f"{job_i:20s} {title_i!s:25s}\n"
+                else:
+                    if job_l == 1:
+                        out += f"{' ':36s} {job_i:20s} {title_i!s:25s}\n"
                     else:
-                        print(f"{' ':20s} {' ':15s} {j_role:20s} {title!s}")
-                    t_count += 1
+                        out += f"{' ':57s} {title_i!s:25s}\n"
+                job_l += 1
+                name_l += 1
+        return out
+
+    def __lt__(self, other):
+        return self.name < other.name
+
+    def __gt__(self, other):
+        return self.name > other.name
+
+    def __eq__(self, other):
+        return self.name == other.name
+
+
+def extract_from_list(in_dict, in_crew_job, in_job_title, in_movie_title):
+    '''
+    Extract the name of a crew member from a job array.
+    '''
+    for name_i in in_crew_job:
+        if name_i in in_dict:
+            in_dict[name_i].add_job_title(in_job_title, in_movie_title)
+        else:
+            in_dict[name_i] = NameJobTitleMap(name_i, in_job_title,
+                                              in_movie_title)
+
+
+def extract_role_from_list(in_dict, in_crew_roles,
+                           in_job_title, in_movie_title):
+    '''
+    Extract the name of the actor from a role object.
+    '''
+    for name_i in in_crew_roles:
+        if name_i.actor in in_dict:
+            in_dict[name_i.actor].add_job_title(in_job_title, in_movie_title)
+        else:
+            in_dict[name_i.actor] = NameJobTitleMap(name_i.actor,
+                                                    in_job_title,
+                                                    in_movie_title)
 
 
 def grab_crew_names(movies):
-    '''Extract crew names from a movie'''
-    name_list = NameList()
+    '''
+    Extract name objects from the arrays specific to
+    the job functions.
+    '''
+    nm_dict = {}
     for movie in movies:
         if movie.crew is not None:
             if movie.crew.directors:
-                for dir_n in movie.crew.directors:
-                    name_list.add(dir_n, "Director", movie.title)
+                extract_from_list(nm_dict, movie.crew.directors,
+                                  'Director', movie.title)
             if movie.crew.writers:
-                for writ_n in movie.crew.writers:
-                    name_list.add(writ_n, "Writer", movie.title)
+                extract_from_list(nm_dict, movie.crew.writers,
+                                  'Writer', movie.title)
             if movie.crew.cinemap:
-                for cinemap_n in movie.crew.cinemap:
-                    name_list.add(cinemap_n, "Cinemaphotographer", movie.title)
+                extract_from_list(nm_dict, movie.crew.cinemap,
+                                  'Cinemaphotographer', movie.title)
             if movie.crew.editors:
-                for editor_n in movie.crew.editors:
-                    name_list.add(editor_n, "Editor", movie.title)
+                extract_from_list(nm_dict, movie.crew.editors,
+                                  'Editor', movie.title)
             if movie.crew.cast:
-                for role in movie.crew.cast.cast:
-                    name_list.add(role.actor, "Cast", movie.title)
-    return name_list
+                extract_role_from_list(nm_dict, movie.crew.cast.cast,
+                                       'Cast', movie.title)
+    return list(nm_dict.values())
+
+
+def list_names(in_names):
+    '''
+    Generate output from the list of passed names.
+    '''
+    print(NameJobTitleMap.header())
+    for name_i in sorted(in_names):
+        print(name_i, end='')
 
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Simple movie list.')
     parser.add_argument('--mediapath', help='path of media library')
+    parser.add_argument('--random', type=int, help='show X names')
     args = parser.parse_args()
     mediapath = args.mediapath or os.environ['MEDIAPATH']
     if not mediapath:
         parser.print_help()
     devices = load_media_dev(mediapath)
     all_movies = compile_movies(devices)
-    names = grab_crew_names(all_movies)
-    names.output()
+    all_names = grab_crew_names(all_movies)
+    if args.random:
+        rand_limit = args.random
+        list_names(random_sample_list(all_names, rand_limit))
+    else:
+        list_names(all_names)